requirements/common.txt
dj-database-url==0.5.0
Django==2.1.2
python-decouple==3.1
django-oauth-toolkit==1.2.0
djangorestframework==3.9.0
django-cors-middleware==1.3.1
drf-yasg==1.11.0
swagger_spec_validator==2.4.1
flex==6.13.2
django-filter==2.0.0